Promoted `or` to `NaturalTransformation.or` + tests + tutorial fixes
