+++ core/src/main/scala/cats/arrow/NaturalTransformation.scala
@@ -1,6 +1,8 @@
 package cats
 package arrow
 
+import cats.data.{Xor, Coproduct}
+
 trait NaturalTransformation[F[_], G[_]] extends Serializable { self =>
   def apply[A](fa: F[A]): G[A]
 
@@ -17,5 +19,13 @@ object NaturalTransformation {
   def id[F[_]]: NaturalTransformation[F, F] =
     new NaturalTransformation[F, F] {
       def apply[A](fa: F[A]): F[A] = fa
+    }
+
+  def or[F[_], G[_], H[_]](f: F ~> H, g: G ~> H): Coproduct[F, G, ?] ~> H =
+    new (Coproduct[F, G, ?] ~> H) {
+      def apply[A](fa: Coproduct[F, G, A]): H[A] = fa.run match {
+        case Xor.Left(ff) => f(ff)
+        case Xor.Right(gg) => g(gg)
+      }
     }
 }
